Check Influential Observations in Cox Models

Identifies influential observations in Cox models using dfbeta residuals.

Usage

check_influential_coxph(model)

Arguments

model

A fitted coxph object.

Value

A list containing dfbeta residuals and influential point indices, or NA if computation fails.

Examples

if (requireNamespace("survival", quietly = TRUE)) {
  model <- survival::coxph(
    survival::Surv(time, status) ~ age + sex,
    data = survival::lung
  )
  check_influential_coxph(model)
}
#> $dfbetas
#>              [,1]          [,2]
#> 1   -0.0008754998 -0.0077651404
#> 2   -0.0151046439  0.0242742997
#> 3    0.2078949655  0.1732024564
#> 4   -0.0346412598 -0.0311084236
#> 5    0.0543934746  0.1475295779
#> 6   -0.4236479842  0.1964847244
#> 7    0.0221347686  0.0438486782
#> 8    0.0201398176  0.0301915723
#> 9   -0.0569125512 -0.0332052811
#> 10  -0.0168065299 -0.0372340687
#> 11  -0.0426464984 -0.0383575546
#> 12  -0.0070125230 -0.0529235238
#> 13  -0.0294205055 -0.0876930314
#> 14  -0.0287422461 -0.0451399252
#> 15   0.0458170903  0.0414576645
#> 16   0.0218757754 -0.0363075735
#> 17  -0.0613047200  0.0631424180
#> 18   0.0126231059  0.0882121998
#> 19  -0.0572434618  0.1083248487
#> 20  -0.0531826928 -0.0454885749
#> 21   0.0038924687 -0.0135035319
#> 22  -0.1172420880  0.1039767851
#> 23   0.0871639368  0.0352083758
#> 24   0.0026696385 -0.0003581999
#> 25  -0.0341488276  0.0188287655
#> 26   0.0025904567 -0.0124974490
#> 27   0.0305741704  0.0455735649
#> 28   0.0458901871 -0.0355215625
#> 29  -0.0048864127 -0.0004496996
#> 30   0.0881479491 -0.0397473631
#> 31   0.0053289360 -0.0058501011
#> 32   0.0787949966 -0.0402426019
#> 33   0.0490756318  0.0168873127
#> 34  -0.0213757787  0.1038869884
#> 35  -0.0231239762 -0.0465617666
#> 36  -0.0035420687  0.1014661306
#> 37  -0.0562483986  0.1446645338
#> 38  -0.0778348926 -0.2603686458
#> 39   0.0760309572 -0.0355622608
#> 40   0.0015236892 -0.0734401574
#> 41  -0.0335215415  0.0252644955
#> 42   0.0752520921  0.0963465566
#> 43  -0.0105606588  0.0260958646
#> 44  -0.0211574956  0.0966057379
#> 45  -0.0383891861  0.0565234522
#> 46   0.1072637889  0.1081041474
#> 47   0.0028861218 -0.0086328996
#> 48   0.0102082978  0.0254481581
#> 49  -0.1426620100  0.0772371950
#> 50   0.0617790072 -0.0218076075
#> 51  -0.0761360563 -0.1061119379
#> 52  -0.0026454020 -0.0327516363
#> 53   0.0369630485 -0.0428684349
#> 54  -0.0242226523 -0.0266166921
#> 55   0.0647327020  0.0722862282
#> 56  -0.0126082894 -0.0440404834
#> 57   0.0152878873  0.1218387907
#> 58  -0.0475680850 -0.0422570972
#> 59   0.0280197537 -0.0342195480
#> 60   0.0030892056  0.0413722679
#> 61   0.0046786671 -0.0024822675
#> 62  -0.0845715864 -0.0379162745
#> 63   0.0489690886 -0.0282634211
#> 64   0.0213172387  0.1121493509
#> 65   0.0317508182 -0.0325677722
#> 66   0.0361576264 -0.0407867839
#> 67   0.0247186298  0.0768667711
#> 68  -0.0353599893 -0.2265218052
#> 69  -0.0129704999  0.0154074299
#> 70   0.0089829924 -0.0237044079
#> 71   0.0019670364  0.1745609500
#> 72  -0.0856831888  0.0582510728
#> 73   0.0880847754 -0.0397517929
#> 74  -0.1827465355 -0.0522966979
#> 75  -0.0668689168  0.0741342904
#> 76   0.0117563168  0.0152674469
#> 77   0.0346899918 -0.0262419585
#> 78  -0.0275297573  0.0452928887
#> 79   0.1510647350 -0.0361730441
#> 80   0.0487517363 -0.0278160416
#> 81   0.0781106213  0.1364837700
#> 82  -0.0678228477 -0.0459975892
#> 83   0.0628545878  0.0192092591
#> 84  -0.1467882358  0.0896084110
#> 85   0.3921987054  0.1440528960
#> 86   0.0147445311 -0.0132969469
#> 87   0.0078758793 -0.0282218166
#> 88  -0.0185256480 -0.0383810018
#> 89   0.2060800446 -0.1246778172
#> 90   0.0484446744 -0.0307562833
#> 91   0.0216373033  0.0698383796
#> 92   0.0173446204 -0.0211843129
#> 93   0.0183009078 -0.0404895300
#> 94  -0.0368599648  0.0710465542
#> 95  -0.0733208814 -0.1429867706
#> 96   0.0702265896 -0.0407620199
#> 97   0.0284626773 -0.0298076395
#> 98   0.0026366945 -0.0097100865
#> 99   0.0039153477  0.0224372411
#> 100  0.0524440440  0.0915770885
#> 101  0.0416266024 -0.1043686938
#> 102  0.0091197846  0.0071053976
#> 103 -0.0002144263 -0.0042683852
#> 104 -0.0043449834 -0.0400599953
#> 105 -0.0277286166 -0.0366397189
#> 106  0.0145084738 -0.0344214171
#> 107  0.0674770333 -0.0892881962
#> 108  0.0247900563 -0.0433356937
#> 109  0.0090853326  0.0017870414
#> 110  0.0296280933  0.0288047744
#> 111  0.0441554016 -0.0423248386
#> 112 -0.1287962601 -0.0460274484
#> 113  0.0121562280 -0.0066241970
#> 114  0.0712343152  0.0751236145
#> 115 -0.0059296724  0.0144794322
#> 116  0.1055738446 -0.0387108442
#> 117 -0.0812020635 -0.0360590331
#> 118 -0.0220902173  0.0372133809
#> 119  0.0044364150 -0.0158667738
#> 120 -0.0624876384  0.0170349434
#> 121 -0.0628657297  0.0258872252
#> 122 -0.0209148429  0.0792022706
#> 123 -0.0075953776 -0.0171817410
#> 124  0.0701207813 -0.0407427681
#> 125 -0.0508454193  0.0558326324
#> 126  0.0079486774 -0.0259770909
#> 127 -0.1102796391 -0.0494295559
#> 128  0.0038703106 -0.0448267017
#> 129 -0.1711727266 -0.1533142065
#> 130  0.1050191649 -0.0800326572
#> 131 -0.0199492884  0.0580345731
#> 132 -0.0592058465 -0.0361227397
#> 133 -0.0363781034 -0.0177939016
#> 134  0.0562256417 -0.0840139674
#> 135  0.0064006974 -0.0191310064
#> 136 -0.1118366617 -0.1265911358
#> 137 -0.0159148002  0.0405422838
#> 138 -0.0007616745  0.0025732465
#> 139  0.0244845675  0.0085313245
#> 140 -0.0147029968 -0.0194709386
#> 141  0.0603792235  0.0758959683
#> 142 -0.1343185438  0.0706692732
#> 143  0.0331409584 -0.0175423604
#> 144  0.0210104986  0.1115434034
#> 145  0.0800080998  0.0746026619
#> 146  0.0465566919 -0.0553224510
#> 147 -0.0106924396  0.0046573417
#> 148 -0.0656695895 -0.0417699356
#> 149  0.1540563399 -0.0355514560
#> 150 -0.0184466524  0.0454481806
#> 151  0.0174858876 -0.0215642090
#> 152  0.0448033069  0.0718811551
#> 153  0.0037887626 -0.0570561145
#> 154 -0.0718797724  0.0843511434
#> 155 -0.0596694240 -0.0411105015
#> 156  0.0009023619 -0.0087256980
#> 157 -0.0338639937 -0.0817752271
#> 158 -0.0039905397 -0.0190594228
#> 159 -0.0037524162 -0.0336198801
#> 160  0.0377005381 -0.0652928187
#> 161  0.0050439724 -0.0743470177
#> 162 -0.1274910512  0.0723516881
#> 163 -0.0400484672  0.0415242380
#> 164 -0.0049335619 -0.0386742437
#> 165 -0.0017265776  0.0657403914
#> 166  0.0180692274 -0.0380182806
#> 167 -0.0195850070  0.0810408535
#> 168 -0.0554017868 -0.0288295369
#> 169 -0.0154350622 -0.0348840403
#> 170  0.0033265544 -0.0180212226
#> 171  0.0196371111  0.0407421528
#> 172 -0.0227381043  0.0462939524
#> 173 -0.0444951727 -0.0368918958
#> 174  0.0877513735 -0.0491446847
#> 175  0.0670122254  0.0389951378
#> 176  0.0151203494 -0.0484428754
#> 177 -0.0457361284 -0.0381564726
#> 178 -0.0550243777  0.0829329268
#> 179 -0.0774037004  0.0955363217
#> 180 -0.0490446149 -0.0575013477
#> 181  0.0256982049  0.0396621551
#> 182  0.1054331475  0.0296822205
#> 183  0.0095735842  0.0393008492
#> 184 -0.0083499398 -0.0482534636
#> 185  0.0401727103 -0.0400825826
#> 186  0.0640902578 -0.0407626992
#> 187  0.0430825977  0.0557402466
#> 188  0.0307088341  0.0312717419
#> 189  0.0029886091 -0.0389437240
#> 190 -0.0372148068 -0.0262726895
#> 191  0.0147785308 -0.0133100357
#> 192 -0.0946840146 -0.0487042017
#> 193 -0.1014422048 -0.0463880512
#> 194 -0.0011749737  0.0286557913
#> 195  0.0128367448 -0.0150137935
#> 196 -0.0358670054  0.0298369374
#> 197 -0.0006935948 -0.0418443066
#> 198 -0.0011450116  0.0351710877
#> 199  0.0331572785 -0.0364671133
#> 200 -0.0260696459 -0.0404507988
#> 201  0.0122766924  0.1090808574
#> 202  0.0783352171 -0.0402117402
#> 203 -0.0006935948 -0.0418443066
#> 204  0.0276916069 -0.0258747675
#> 205 -0.0006853915 -0.0404931644
#> 206 -0.0123964622 -0.0413747825
#> 207  0.0207581037 -0.0324451520
#> 208 -0.0853313363  0.0722277057
#> 209 -0.0310899565  0.0313906156
#> 210  0.0066960722 -0.0225382574
#> 211  0.0091757150 -0.0405552410
#> 212 -0.0171069308  0.0281933985
#> 213 -0.0228432841  0.0222138072
#> 214 -0.0010589968 -0.0114163055
#> 215  0.0084963758 -0.0443938471
#> 216 -0.0063899987  0.0283235419
#> 217  0.0363039606 -0.0191654419
#> 218  0.0518846353 -0.0242752926
#> 219 -0.0233159119 -0.0418493952
#> 220  0.0101534822 -0.0220336137
#> 221 -0.0147379976  0.0237885782
#> 222 -0.0256393297 -0.0406065859
#> 223  0.0850556896 -0.0321191505
#> 224 -0.0671666817  0.0237540471
#> 225  0.0667138241  0.0183357585
#> 226 -0.0150012447 -0.0166388944
#> 227 -0.0076796615  0.0176549749
#> 228  0.0088773525 -0.0234292606
#> 
#> $influential_points
#>  3  6 38 68 85 89 
#>  3  6 38 68 85 89 
#>